Fixed compile error, moved linux-only test to its own test file
